Refactor login functionality and validation
