+++ lib/core/Networking/api_services.dart
@@ -10,4 +10,4 @@ abstract class ApiServices{
 
   @POST(ApiConstants.login)
   Future<LoginRequestResponse> login(@Body() LoginRequestBody loginRequestBody);
-}
+}
