+++ lib/core/Widgets/app_text_form_field.dart
@@ -33,6 +33,7 @@ class AppTextFormField extends StatefulWidget {
   });
 
   @override
+  // ignore: library_private_types_in_public_api
   _AppTextFormFieldState createState() => _AppTextFormFieldState();
 }
