+++ lib/features/login/ui/login_bloc_listener.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_adv_course_2/core/Helpers/extensions.dart';
+import 'package:flutter_adv_course_2/core/Helpers/functions.dart';
 import 'package:flutter_adv_course_2/core/Routing/routes.dart';
 import 'package:flutter_adv_course_2/core/Theming/colors.dart';
-import 'package:flutter_adv_course_2/core/Theming/styles.dart';
 import 'package:flutter_adv_course_2/features/login/logic/cubit/login_cubit.dart';
 import 'package:flutter_adv_course_2/features/login/logic/cubit/login_state.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -42,31 +42,4 @@ class LoginBlocListener extends StatelessWidget {
     );
   }
 
-  void setupMethodState(BuildContext context, String error) {
-        context.pop();
-    showDialog(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          icon: const Icon(
-            Icons.error,
-            color: Colors.red,
-            size: 32,
-          ),
-          content: Text(
-            error,
-            style: TextStyles.font15DarkBlueMedium,
-          ),
-          actions: [
-            TextButton(
-                onPressed: () {
-                  context.pop();
-                },
-                child: Text('Got it',style: TextStyles.font14BlueSemiBold,))
-          ],
-        );
-      },
-    );
-            
-  }
 }
