+++ lib/features/login/ui/login_screen.dart
@@ -1,14 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_adv_course_2/core/Helpers/functions.dart';
 import 'package:flutter_adv_course_2/core/Helpers/spacing.dart';
 import 'package:flutter_adv_course_2/core/Theming/styles.dart';
 import 'package:flutter_adv_course_2/core/Widgets/app_text_button.dart';
-import 'package:flutter_adv_course_2/features/login/data/models/login_request_body.dart';
-import 'package:flutter_adv_course_2/features/login/logic/cubit/login_cubit.dart';
 import 'package:flutter_adv_course_2/features/login/ui/login_bloc_listener.dart';
 import 'package:flutter_adv_course_2/features/login/ui/widgets/already_have_account_text.dart';
 import 'package:flutter_adv_course_2/features/login/ui/widgets/email_and_password.dart';
 import 'package:flutter_adv_course_2/features/login/ui/widgets/terms_and_conditions.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class LoginScreen extends StatelessWidget {
@@ -68,15 +66,5 @@ class LoginScreen extends StatelessWidget {
     );
   }
 
-  void validateTheLogin(BuildContext context) {
-    final loginCubit = context.read<LoginCubit>();
-    if (loginCubit.formKey.currentState!.validate()) {
-      loginCubit.emitLoginState(
-        LoginRequestBody(
-          email: loginCubit.emailController.text,
-          password: loginCubit.passwordController.text,
-        ),
-      );
-    }
-  }
+
 }
